Move SI-10481 test into the junit project
