+++ test/junit/src/test/scala/strawman/collection/concurrent/TrieMapTest.scala
@@ -4,6 +4,8 @@ import strawman.collection.{IterableOnce, View}
 import strawman.collection.immutable.{List, Nil}
 import org.junit.{Assert, Test}
 
+import scala.util.hashing.Hashing
+
 class TrieMapTest {
 
   private def check[T](result2: List[Any])(f: TrieMap[String, String] => IterableOnce[Any]) = {
@@ -52,5 +54,19 @@ class TrieMapTest {
   @Test
   def mapValues(): Unit = {
     check(List(("k", "v")))(_.mapValues(x => x))
+  }
+
+  @Test
+  def customHashingAndEquiv_10481(): Unit = {
+    val h = new Hashing[Int] { def hash(i: Int) = i % 4 }
+    val e = new Equiv[Int] { def equiv(x: Int, y: Int) = (x % 8) == (y % 8) }
+    val xs = new TrieMap[Int, String](h, e)
+    xs.put(0, "zero")
+    Assert.assertEquals(Some("zero"), xs.get(0))
+    Assert.assertEquals(Some("zero"), xs.get(8)) // 8 and 0 are equivalent keys according to our custom equiv
+    xs.put(4, "four") // 4 and 0 have the same hash according to our custom hashing, but they
+    // are different keys (collision)
+    Assert.assertEquals(Some("zero"), xs.get(8))
+    Assert.assertEquals(Some("four"), xs.get(4))
   }
 }
